% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckResults.R
\name{checkBestPaths}
\alias{checkBestPaths}
\title{This function checks the results returned by GetBestPaths}
\usage{
checkBestPaths(dataset, BestPaths, pathLength, threshold_percent, nCases,
  nControls, method)
}
\description{
This function checks the results returned by GetBestPaths
}
