% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessPaths.R
\name{GetBestPaths}
\alias{GetBestPaths}
\title{This function runs the causal engine for a given phenotype dataset.}
\usage{
GetBestPaths(dataset, nCases, nControls, method = 1,
       threshold_percent = 0.05, K = 10, pathLength = 5, iterations = 100,
       strataF = NA)
}
\arguments{
\item{dataset}{Filename of the phenotype data. First column must correspond
to the gene symbols; all other columns correspond to the patients.
First column must be named symbols and must contrain strings corresponding
to the symbols of the genes. All other columns either contain a 0 if the
gene is not present in the patient, and 1 or 2 if the gene is present
in the patient.}

\item{nCases}{The number of cases (Must be greater than 1).}

\item{nControls}{The number of controls (Must be greater than 1).}

\item{method}{The method to be invoked. The value 1 calls the unsigned method
while the value 2 calls the signed method.}

\item{threshold_percent}{All genes which occur in less than emph{threshold_percent}
of patients are kept for processing. The default
value is 0.05.}

\item{K}{The top K paths to be returned of each path length. The default
value is 10.}

\item{pathLength}{The maximum path length for which the top K paths
will be computed. The default value is 5 (The maximum
value for \emph{pathLength} is 5).}

\item{iterations}{The number of iterations to be used in computing the
p-values. The default value is 100 (The minimum acceptable
value is 1).}

\item{strataF}{Filename for the strata file. \emph{strataF} must be a file
of two columns, first column named \emph{subjid }contains
the patient names similar to the dataset columns, and the
second column named \emph{stratum} must contain the integer
corresponding to the stratum.}

\item{nthreads}{The number of threads to be used in the parallel region
of the code. Default value is \emph{nthreads = NA}, in
which case \emph{nthreads} will be set to the maximum number of
threads that can be created. If \emph{nthreads} is greater
than the maximum number of threads that can be created, then
\emph{nthreads} will default back to the maximum number of threads
that can be created.}

\item{path}{The path where the package will write and read from files
created by the package.}
}
\value{
This function returns a list with the following items:
        \item{SignedPaths}{The top K signed paths for each length.}
        \item{Paths}{The top K paths for each length.}
        \item{Lengths}{The lengths of each path.}
        \item{Scores}{The scores of each path.}
        \item{Pvalues}{The p-values of each path.}
        \item{Cases}{The number of cases for each path.}
        \item{Controls}{The number of controls for each path.}
}
\description{
This function runs the causal engine for a given phenotype dataset.
}
\references{
Franceschini, A (2013). STRING v9.1: protein-protein interaction networks, with increased coverage
            and integration. In:'Nucleic Acids Res. 2013 Jan;41(Database issue):D808-15. doi: 10.1093/nar/gks1094.
            Epub 2012 Nov 29'.
}
\author{
Carl Tony Fakhry
}
