% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessPaths.R
\name{GWASPA}
\alias{GWASPA}
\title{This function performs genome-wide association study pathway analysis (GWASPA) to
identify statistically significant associations between variants on a gene regulatory pathways
and a given phenotype.}
\usage{
GWASPA(dataset, nCases, nControls, Signed.GWASPA = FALSE, Decorated.Pvalues = TRUE,
       threshold = 0.05, K = 10, pathLength = 5, n_permutations = 100,
       strataF = NA, nthreads = NA)
}
\arguments{
\item{dataset}{Filename of the phenotype data. First column must correspond
to the gene symbols, and all all other columns correspond to the
patients. The First column must be named \code{Gene.Symbols} and
must contain strings corresponding to the symbols of the genes.
All other columns either contain a 1 if the patient has a variant
in the gene, and 0 otherwise.}

\item{nCases}{The number of cases (Must be greater than 1).}

\item{nControls}{The number of controls (Must be greater than 1).}

\item{Signed.GWASPA}{A boolean parameter indicating whether to compute the
signed-GWASPA. The default value is \code{Signed.GWASPA = FALSE}.}

\item{Decorated.Pvalues}{A boolean parameter indicating whether to compute
decorated P-values. The default value is
\code{Decorated.Pvalues = TRUE}.}

\item{threshold}{All variants which occur in more than \code{threshold * 100}
percentage of patients are kept for processing. The default
value is \code{threshold = 0.05}.}

\item{K}{The top \code{K} paths to be returned of each path length. The default
value is \code{K = 10}.}

\item{pathLength}{The maximum path length for which the top \code{K} paths
will be computed. The default value is \code{pathLength = 5}.
Note that \code{pathLength} cannot be set to less than 1 or
greater than 5.}

\item{n_permutations}{The number of permutations to be used for computing the
p-values. The default value is 100 (The minimum acceptable
value is 1).}

\item{strataF}{Filename for the strata file. \code{strataF} must be a file
of two columns: The first column must be named \code{subjid}
and contains the patient names and must match the patients names
in \code{dataset} file, and the
second column named \code{stratum} must contain the integer
corresponding to the stratum.}

\item{nthreads}{The number of threads to be used in the parallel region
of the code. Default value is \code{nthreads = NA}, in
which case \code{nthreads} only one thread will be used.}
}
\value{
This function returns a list  with the following items:
        \item{GWASPA.Results}{The top \code{K} paths for each length sorted in
         increasing order of the p-values. The results are stored in a data
         frame with the following columns:\cr
         \code{-} \code{SignedPaths  } The top \code{K} signed paths for each length.\cr
         \code{-} \code{Paths  } The top \code{K} paths for each length.\cr
         \code{-} \code{Lengths  } The length of each path.\cr
         \code{-} \code{Scores  } The score of each path.\cr
         \code{-} \code{Pvalues  } The p-value of each path.\cr
         \code{-} \code{Cases  } The number of cases for each path.\cr
         \code{-} \code{Controls  } The number of controls for each path.\cr}
        \item{Decorated.Pvalues.Results}{If \code{Decorated.Pvalues = TRUE} then the decorated p-values
         of the partitioned paths are computed. The results are stored in a data frame with the following columns:\cr
         \code{-} \code{SignedPaths  } The top \code{K} signed paths for each length.\cr
         \code{-} \code{Paths  } The top \code{K} paths for each length.\cr
         \code{-} \code{Subpaths1  } The first subpath partition.\cr
         \code{-} \code{Subpaths1.Cases  } The number of cases for \code{Subpaths1}.\cr
         \code{-} \code{Subpaths1.Controls  } The number of controls for \code{Subpaths1}.\cr
         \code{-} \code{Subpaths2  } The second subpath partition.\cr
         \code{-} \code{Subpaths2.Cases  } The number of cases for \code{Subpaths2}.\cr
         \code{-} \code{Subpaths2.Controls  } The number of controls for \code{Subpaths2}.\cr
         \code{-} \code{Direction  } Direction of splitting along the path.\cr
         \code{-} \code{Lengths  } The length of each path.\cr
         \code{-} \code{Scores  } The score of each path.\cr
         \code{-} \code{Pvalues  } The p-value of each path.\cr
         \code{-} \code{DecoratedPvalues  } The decorated p-value of each path.\cr
         \code{-} \code{Cases  } The number of cases for each path.\cr
         \code{-} \code{Controls  } The number of controls for each path.\cr}
}
\description{
This function performs genome-wide association study pathway analysis (GWASPA) to
             identify statistically significant associations between variants on a gene
             regulatory pathways and a given phenotype.
}
\references{
Franceschini, A (2013). STRING v9.1: protein-protein interaction networks, with increased coverage
            and integration. In:'Nucleic Acids Res. 2013 Jan;41(Database issue):D808-15. doi: 10.1093/nar/gks1094.
            Epub 2012 Nov 29'.
}
\author{
Carl Tony Fakhry
}
