% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecoratedPvalue.R
\name{getDecoratedPvalues}
\alias{getDecoratedPvalues}
\title{This function computes the decorated p-values for each path.}
\usage{
getDecoratedPvalues(dataset, BestPaths, pathLength, nCases,
                                    nControls, method, n_permutations = 100, strataF = NA)
}
\arguments{
\item{dataset}{Filename of the phenotype data. First column must correspond
to the gene symbols; all other columns correspond to the patients.
First column must be named symbols and must contrain strings corresponding
to the symbols of the genes. All other columns either contain a 0 if the
gene is not present in the patient, and 1 or 2 if the gene is present
in the patient.}

\item{BestPaths}{A data frame returned from calling \link{ProcessPaths}.}

\item{pathLength}{The maximum path length for which the top K paths
will be computed. The default value is 5 (The maximum
value for \emph{pathLength} is 5).}

\item{nCases}{The number of cases (Must be greater than 1).}

\item{nControls}{The number of controls (Must be greater than 1).}

\item{method}{The method to be invoked. The value 1 calls the unsigned method
while the value 2 calls the signed method.}

\item{n_permutations}{The number of permutations to be used in computing the
p-values. The default value is 100.}

\item{strataF}{Filename for the strata file. \emph{strataF} must be a file
of two columns, first column named \emph{subjid }contains
the patient names similar to the dataset columns, and the
second column named \emph{stratum} must contain the integer
corresponding to the stratum.}
}
\value{
This function returns a list with the following items:
        \item{SignedPaths}{The top K signed paths for each length.}
        \item{Paths}{The top K paths for each length.}
        \item{Subpaths1}{The first subpath.}
        \item{Subpaths2}{The second subpath.}
        \item{Direction}{Direction of splitting along the path.}
        \item{Lengths}{The lengths of each path.}
        \item{Scores}{The scores of each path.}
        \item{Pvalues}{The p-values of each path.}
        \item{DecoratedPvalues}{The decorated p-values for the second subpath.}
}
\description{
This function computes the decorated p-values for each path.
}
\references{
Franceschini, A (2013). STRING v9.1: protein-protein interaction networks, with increased coverage
            and integration. In:'Nucleic Acids Res. 2013 Jan;41(Database issue):D808-15. doi: 10.1093/nar/gks1094.
            Epub 2012 Nov 29'.
}
\author{
Carl Tony Fakhry
}
